PPTX-AGENTS/backend/requirements.txt
# PPTX Agents — Backend Python Dependencies

# Web framework & SSE
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sse-starlette>=2.2.1

# Azure AI
azure-ai-projects>=1.0.0b1
azure-identity>=1.19.0

# PPTX read/write
python-pptx>=1.0.2

# Markdown extraction (PPTX → text)
markitdown[pptx]>=0.1.0

# Image rendering (best-effort slide → PNG)
Pillow>=11.0.0
# pdf2image>=1.17.0   # uncomment if LibreOffice is available for high-fidelity rendering

# Utilities
python-dotenv>=1.0.0
pydantic>=2.10.0
